deleteUser and updateUserNickname

+++ src/main/java/kr/io/classicgame/service/UserServiceImpl.java
@@ -24,12 +24,12 @@ public class UserServiceImpl implements UserService {
 			return null;
 		}
 	}
-	
+
 	@Transactional
 	public boolean insertUser(User user) {
 		// null인 경우 @ExceptionHandler를 통해 예외처리 필요. 닉네임, mail도 검증 필요.
 		boolean result = false;
-		
+
 		if (getUser(user) == null) {
 			userRepo.save(user);
 			result = true;
@@ -37,14 +37,29 @@ public class UserServiceImpl implements UserService {
 		return result;
 	}
 
-	public void deleteUser(User user) {
-		userRepo.deleteById(user.getId());
+	@Transactional
+	public boolean deleteUser(User user) {
+
+		boolean result = false;
+
+		if (getUser(user) != null) {
+			userRepo.deleteById(user.getId());
+			result = true;
+		}
+		return result;
 	}
 
-	public void updateUserNickname(User user) {
-		User findUser = userRepo.findById(user.getId()).get();
-		findUser.setNickname(user.getNickname());
-		userRepo.save(findUser);
-	}
+	@Transactional
+	public boolean updateUserNickname(User user) {
 
+		boolean result = false;
+
+		if (getUser(user) != null) {
+			User findUser = userRepo.findById(user.getId()).get();
+			findUser.setNickname(user.getNickname());
+			userRepo.save(findUser);
+			result = true;
+		}
+		return result;
+	}
 }
